Moved files from app into activity/controller/model folders
