+++ app/javadocs/com/example/nidoqueue/DataCalc.html
@@ -98,7 +98,7 @@ var activeTableTab = "activeTableTab";
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>com.example.nidoqueue.DataCalc</li>
+<li>com.example.nidoqueue.model.DataCalc</li>
 </ul>
 </li>
 </ul>
