+++ app/javadocs/com/example/nidoqueue/ExpListAdapter.html
@@ -98,10 +98,10 @@ var activeTableTab = "activeTableTab";
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>androidx.recyclerview.widget.RecyclerView.Adapter&lt;com.example.nidoqueue.Holder&gt;</li>
+<li>androidx.recyclerview.widget.RecyclerView.Adapter&lt;com.example.nidoqueue.activity.Holder&gt;</li>
 <li>
 <ul class="inheritance">
-<li>com.example.nidoqueue.ExpListAdapter</li>
+<li>com.example.nidoqueue.activity.ExpListAdapter</li>
 </ul>
 </li>
 </ul>
@@ -113,7 +113,7 @@ var activeTableTab = "activeTableTab";
 <hr>
 <br>
 <pre>public class <span class="typeNameLabel">ExpListAdapter</span>
-extends androidx.recyclerview.widget.RecyclerView.Adapter&lt;com.example.nidoqueue.Holder&gt;</pre>
+extends androidx.recyclerview.widget.RecyclerView.Adapter&lt;com.example.nidoqueue.activity.Holder&gt;</pre>
 <div class="block">Classname:   ExpListAdapter.java
  Version:     Prototype
  Date:        March 19th, 2021
@@ -143,11 +143,11 @@ extends androidx.recyclerview.widget.RecyclerView.Adapter&lt;com.example.nidoque
 </tr>
 <tr id="i1" class="rowColor">
 <td class="colFirst"><code>void</code></td>
-<td class="colLast"><code><span class="memberNameLink"><a href="../../../com/example/nidoqueue/ExpListAdapter.html#onBindViewHolder-com.example.nidoqueue.Holder-int-">onBindViewHolder</a></span>(com.example.nidoqueue.Holder&nbsp;holder,
+<td class="colLast"><code><span class="memberNameLink"><a href="../../../com/example/nidoqueue/ExpListAdapter.html#onBindViewHolder-com.example.nidoqueue.activity.Holder-int-">onBindViewHolder</a></span>(com.example.nidoqueue.activity.Holder&nbsp;holder,
                 int&nbsp;position)</code>&nbsp;</td>
 </tr>
 <tr id="i2" class="altColor">
-<td class="colFirst"><code>com.example.nidoqueue.Holder</code></td>
+<td class="colFirst"><code>com.example.nidoqueue.activity.Holder</code></td>
 <td class="colLast"><code><span class="memberNameLink"><a href="../../../com/example/nidoqueue/ExpListAdapter.html#onCreateViewHolder-android.view.ViewGroup-int-">onCreateViewHolder</a></span>(android.view.ViewGroup&nbsp;parent,
                   int&nbsp;viewType)</code>&nbsp;</td>
 </tr>
@@ -187,27 +187,27 @@ extends androidx.recyclerview.widget.RecyclerView.Adapter&lt;com.example.nidoque
 <li class="blockList">
 <h4>onCreateViewHolder</h4>
 <pre>@NonNull
-public&nbsp;com.example.nidoqueue.Holder&nbsp;onCreateViewHolder(@NonNull
+public&nbsp;com.example.nidoqueue.activity.Holder&nbsp;onCreateViewHolder(@NonNull
                                                                  android.view.ViewGroup&nbsp;parent,
                                                                  int&nbsp;viewType)</pre>
 <dl>
 <dt><span class="overrideSpecifyLabel">Specified by:</span></dt>
-<dd><code>onCreateViewHolder</code>&nbsp;in class&nbsp;<code>androidx.recyclerview.widget.RecyclerView.Adapter&lt;com.example.nidoqueue.Holder&gt;</code></dd>
+<dd><code>onCreateViewHolder</code>&nbsp;in class&nbsp;<code>androidx.recyclerview.widget.RecyclerView.Adapter&lt;com.example.nidoqueue.activity.Holder&gt;</code></dd>
 </dl>
 </li>
 </ul>
-<a name="onBindViewHolder-com.example.nidoqueue.Holder-int-">
+<a name="onBindViewHolder-com.example.nidoqueue.activity.Holder-int-">
 <!--   -->
 </a>
 <ul class="blockList">
 <li class="blockList">
 <h4>onBindViewHolder</h4>
 <pre>public&nbsp;void&nbsp;onBindViewHolder(@NonNull
-                             com.example.nidoqueue.Holder&nbsp;holder,
+                             com.example.nidoqueue.activity.Holder&nbsp;holder,
                              int&nbsp;position)</pre>
 <dl>
 <dt><span class="overrideSpecifyLabel">Specified by:</span></dt>
-<dd><code>onBindViewHolder</code>&nbsp;in class&nbsp;<code>androidx.recyclerview.widget.RecyclerView.Adapter&lt;com.example.nidoqueue.Holder&gt;</code></dd>
+<dd><code>onBindViewHolder</code>&nbsp;in class&nbsp;<code>androidx.recyclerview.widget.RecyclerView.Adapter&lt;com.example.nidoqueue.activity.Holder&gt;</code></dd>
 </dl>
 </li>
 </ul>
@@ -220,7 +220,7 @@ public&nbsp;com.example.nidoqueue.Holder&nbsp;onCreateViewHolder(@NonNull
 <pre>public&nbsp;int&nbsp;getItemCount()</pre>
 <dl>
 <dt><span class="overrideSpecifyLabel">Specified by:</span></dt>
-<dd><code>getItemCount</code>&nbsp;in class&nbsp;<code>androidx.recyclerview.widget.RecyclerView.Adapter&lt;com.example.nidoqueue.Holder&gt;</code></dd>
+<dd><code>getItemCount</code>&nbsp;in class&nbsp;<code>androidx.recyclerview.widget.RecyclerView.Adapter&lt;com.example.nidoqueue.activity.Holder&gt;</code></dd>
 </dl>
 </li>
 </ul>
