+++ app/javadocs/com/example/nidoqueue/ExperimentManager.html
@@ -92,7 +92,7 @@
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>com.example.nidoqueue.ExperimentManager</li>
+<li>com.example.nidoqueue.controller.ContextManager.ExperimentManager</li>
 </ul>
 </li>
 </ul>
