+++ app/javadocs/com/example/nidoqueue/MainActivity.html
@@ -125,7 +125,7 @@ var activeTableTab = "activeTableTab";
 <li><a href="../../../com/example/nidoqueue/AbstractActivity.html" title="class in com.example.nidoqueue">com.example.nidoqueue.AbstractActivity</a></li>
 <li>
 <ul class="inheritance">
-<li>com.example.nidoqueue.MainActivity</li>
+<li>com.example.nidoqueue.activity.MainActivity</li>
 </ul>
 </li>
 </ul>
