+++ app/javadocs/com/example/nidoqueue/Trial.html
@@ -98,7 +98,7 @@ var activeTableTab = "activeTableTab";
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>com.example.nidoqueue.Trial</li>
+<li>com.example.nidoqueue.model.Trial</li>
 </ul>
 </li>
 </ul>
