+++ app/javadocs/index-files/index-11.html
@@ -82,7 +82,7 @@
 <dd>&nbsp;</dd>
 <dt><span class="memberNameLink"><a href="../com/example/nidoqueue/WelcomeActivity.html#onBackPressed--">onBackPressed()</a></span> - Method in class com.example.nidoqueue.<a href="../com/example/nidoqueue/WelcomeActivity.html" title="class in com.example.nidoqueue">WelcomeActivity</a></dt>
 <dd>&nbsp;</dd>
-<dt><span class="memberNameLink"><a href="../com/example/nidoqueue/ExpListAdapter.html#onBindViewHolder-com.example.nidoqueue.Holder-int-">onBindViewHolder(Holder, int)</a></span> - Method in class com.example.nidoqueue.<a href="../com/example/nidoqueue/ExpListAdapter.html" title="class in com.example.nidoqueue">ExpListAdapter</a></dt>
+<dt><span class="memberNameLink"><a href="../com/example/nidoqueue/ExpListAdapter.html#onBindViewHolder-com.example.nidoqueue.activity.Holder-int-">onBindViewHolder(Holder, int)</a></span> - Method in class com.example.nidoqueue.<a href="../com/example/nidoqueue/ExpListAdapter.html" title="class in com.example.nidoqueue">ExpListAdapter</a></dt>
 <dd>&nbsp;</dd>
 <dt><span class="memberNameLink"><a href="../com/example/nidoqueue/DatabaseManager.html#onCreate--">onCreate()</a></span> - Method in class com.example.nidoqueue.<a href="../com/example/nidoqueue/DatabaseManager.html" title="class in com.example.nidoqueue">DatabaseManager</a></dt>
 <dd>&nbsp;</dd>
