--- app/src/main/java/com/example/nidoqueue/ExperimentManager.java
@@ -1,11 +1,0 @@
-package com.example.nidoqueue;
-
-/**
- * Classname:   ExperimentManager.java
- * Version:     Prototype
- * Date:        March 19th, 2021
- * Purpose:     This handles experiment transitions similar to the UserControl class.
- * Issues:      Incomplete, will be worked on in the future.
- */
-public class ExperimentManager {
-}
