+++ app/src/main/java/com/example/nidoqueue/controller/ContextManager.java
@@ -26,4 +26,13 @@ public class ContextManager {
 
     public AbstractActivity getActivity() { return context; }
 
+    /**
+     * Classname:   ExperimentManager.java
+     * Version:     Prototype
+     * Date:        March 19th, 2021
+     * Purpose:     This handles experiment transitions similar to the UserControl class.
+     * Issues:      Incomplete, will be worked on in the future.
+     */
+    public static class ExperimentManager {
+    }
 }
